Move ZstdCompressor to a package extension
