+++ Project.toml
@@ -20,6 +20,12 @@ Pkg = "44cfe95a-1eb2-52ea-b672-e2afdf69b78f"
 URIs = "5c2747f8-b7ea-4ff2-ba2e-563bfd36b1d4"
 ZipArchives = "49080126-0e18-4c2a-b176-c102e4b3760c"
 
+[weakdeps]
+ChunkCodecLibZstd = "55437552-ac27-4d47-9aa3-63184e8fd398"
+
+[extensions]
+ChunkCodecLibZstdExt = "ChunkCodecLibZstd"
+
 [compat]
 AWSS3 = "0.10, 0.11"
 Blosc = "0.5, 0.6, 0.7"
