+++ test/Project.toml
@@ -1,4 +1,5 @@
 [deps]
+ChunkCodecLibZstd = "55437552-ac27-4d47-9aa3-63184e8fd398"
 CondaPkg = "992eb4ea-22a4-4c89-a5bb-47a3300528ab"
 Dates = "ade2ca70-3891-5945-98fb-dc099432e06a"
 JSON = "682c06a0-de6a-54ab-a142-c8b1cf79cde6"
